+++ translator/src/main/scala/soda/translator/language/Translation.scala
@@ -10,13 +10,17 @@ case class Translation () {
 
   lazy val SodaColon: String = ":"
 
+  lazy val SodaInReservedWord: String = "in"
+  lazy val SodaInPattern: String = SodaInReservedWord + " "
+  lazy val ScalaInTranslation: String = " } "
+
   lazy val SodaClassReservedWord: String = "class"
 
   lazy val ScalaDefinition: String = "def"
   lazy val ScalaValue: String = "lazy val"
   lazy val ScalaEntryPoint: String = "object EntryPoint {\n  def main(args: Array[String]): Unit = Main().main(args)\n}\n"
 
-  lazy val SodaReservedWords = Seq ("=", ":", "->", ":=", "if", "then", "else", "class", "has", "extends", "with", "this", "subtype", "supertype", "false", "true", "not", "and", "or", "package", "import", "is", "*", "-", "+"
+  lazy val SodaReservedWords = Seq ("=", ":", "->", ":=", "if", "then", "else", "let", "in", "class", "has", "extends", "with", "this", "subtype", "supertype", "false", "true", "not", "and", "or", "package", "import", "is", "*", "-", "+"
   )
 
 /**
@@ -51,7 +55,7 @@ case class Translation () {
   lazy val Synonym: Seq [(String, String )] = Seq (("is", "=")
   )
 
-  lazy val MainTranslation: Seq [(String, String )] = Seq ((":", ":"), ("->", "=>"), (":=", "="), ("if", "if ("), ("then", ")"), ("else", "else"), ("extends", "extends"), ("with", "with"), ("this", "this"), ("subtype", "<:"), ("supertype", ">:"), ("false", "false"), ("true", "true"), ("not", "!"), ("and", "&&"), ("or", "||"),
+  lazy val MainTranslation: Seq [(String, String )] = Seq ((":", ":"), ("->", "=>"), (":=", "="), ("if", "if ("), ("then", ")"), ("else", "else"), ("let", "{"), ("in", " "), ("extends", "extends"), ("with", "with"), ("this", "this"), ("subtype", "<:"), ("supertype", ">:"), ("false", "false"), ("true", "true"), ("not", "!"), ("and", "&&"), ("or", "||"),
     /** Annotations */
     ("@new", "new")
   )